correct the blending modalities and add alpha

+++ phm/data/data.py
@@ -10,7 +10,7 @@ from functools import lru_cache
 from dataclasses import dataclass
 from typing import Any, Dict, List, Tuple, Union
 
-modal_to_image = lambda img : ((img - np.min(img)) / (np.max(img) - np.min(img))) * 255.0
+modal_to_image = lambda img : (((img - np.min(img)) / (np.max(img) - np.min(img))) * 255.0).astype(np.uint8)
 gray_to_rgb = lambda img : np.stack((img, img, img), axis=2)
 
 __modality_loaders = {}
@@ -35,13 +35,11 @@ def load_entity(file_type : str, file : str):
         raise ValueError(f'{file_type} loader does not exist!')
 
     return __modality_loaders[file_type](file, file_type)
-
 @dataclass
 class MMERecord:
     data : Any
     file : str
     type : str
-
 class MMEContainer(object):
     def __init__(self, 
         cid : str = '',
